chore(backend): tidy up blended data mart schemas, fixtures, and report column config
Split relationship-schemas.ts into per-concept files (join-condition,
aggregate-function, report-column-config) and rename blended-fields-config
to singular .schema.ts. Drop the unused JoinConditionsUpdateSchema.

Wire ReportColumnConfigSchema as a Zod transformer on Report.columnConfig,
and propagate the ReportColumnConfig type alias across the report DTOs and
commands so the schema is the single source of truth.

Extract spec fixtures for blended query builders into __fixtures__ — all
six builder specs (5 dialects + the abstract base) now share makeChain /
makeRelationship / createBuildContext.

Move logBlendedSqlIfNeeded into report-run-logging as a pure helper so
BlendedReportDataService no longer owns logging concerns, and guard
CopyReportAsDataMartService against an empty SQL returned by the composer.

Align argument order in relationship commands (projectId before userId)
to match list commands and the Update*Command convention.
