+++ apps/backend/src/data-marts/data-storage-types/interfaces/blended-query-builder.interface.ts
@@ -1,6 +1,6 @@
 import { DataMartRelationship } from '../../entities/data-mart-relationship.entity';
 import { DataStorageType } from '../enums/data-storage-type.enum';
-import { AggregateFunction } from '../../dto/schemas/relationship-schemas';
+import { AggregateFunction } from '../../dto/schemas/aggregate-function.schema';
 
 export interface BlendedFieldConfig {
   targetFieldName: string;
