+++ apps/backend/src/data-marts/dto/domain/create-relationship.command.ts
@@ -1,4 +1,4 @@
-import { JoinCondition } from '../schemas/relationship-schemas';
+import { JoinCondition } from '../schemas/join-condition.schema';
 
 export class CreateRelationshipCommand {
   constructor(
@@ -6,8 +6,8 @@ export class CreateRelationshipCommand {
     public readonly targetDataMartId: string,
     public readonly targetAlias: string,
     public readonly joinConditions: JoinCondition[],
-    public readonly userId: string,
     public readonly projectId: string,
+    public readonly userId: string,
     public readonly roles: string[]
   ) {}
 }
