+++ apps/backend/src/data-marts/dto/domain/delete-relationship.command.ts
@@ -2,8 +2,8 @@ export class DeleteRelationshipCommand {
   constructor(
     public readonly relationshipId: string,
     public readonly sourceDataMartId: string,
-    public readonly userId: string,
     public readonly projectId: string,
+    public readonly userId: string,
     public readonly roles: string[]
   ) {}
 }
