+++ apps/backend/src/data-marts/dto/domain/get-relationship.command.ts
@@ -2,8 +2,8 @@ export class GetRelationshipCommand {
   constructor(
     public readonly relationshipId: string,
     public readonly sourceDataMartId: string,
-    public readonly userId: string,
     public readonly projectId: string,
+    public readonly userId: string,
     public readonly roles: string[]
   ) {}
 }
