+++ apps/backend/src/data-marts/dto/domain/relationship.dto.ts
@@ -1,4 +1,4 @@
-import { JoinCondition } from '../schemas/relationship-schemas';
+import { JoinCondition } from '../schemas/join-condition.schema';
 import { DataMartStatus } from '../../enums/data-mart-status.enum';
 import { UserProjectionDto } from '../../../idp/dto/domain/user-projection.dto';
