+++ apps/backend/src/data-marts/dto/domain/update-relationship.command.ts
@@ -1,11 +1,11 @@
-import { JoinCondition } from '../schemas/relationship-schemas';
+import { JoinCondition } from '../schemas/join-condition.schema';
 
 export class UpdateRelationshipCommand {
   constructor(
     public readonly relationshipId: string,
     public readonly sourceDataMartId: string,
-    public readonly userId: string,
     public readonly projectId: string,
+    public readonly userId: string,
     public readonly roles: string[],
     public readonly targetAlias?: string,
     public readonly joinConditions?: JoinCondition[]
