+++ apps/backend/src/data-marts/dto/presentation/data-mart-response-api.dto.ts
@@ -8,7 +8,7 @@ import { DataStorageResponseApiDto } from './data-storage-response-api.dto';
 import { DataMartSchema } from '../../data-storage-types/data-mart-schema.type';
 import { ConnectorState as ConnectorStateData } from '../../connector-types/interfaces/connector-state';
 import { ConnectorStateResponseApiDto } from './connector-state-response-api.dto';
-import { BlendedFieldsConfig } from '../schemas/blended-fields-config.schemas';
+import { BlendedFieldsConfig } from '../schemas/blended-fields-config.schema';
 
 export class DataMartResponseApiDto {
   @ApiProperty({ example: '9cabc24e-1234-4a5a-8b12-abcdef123456' })
