+++ apps/backend/src/data-marts/dto/presentation/update-relationship-request-api.dto.ts
@@ -10,7 +10,7 @@ import {
 } from 'class-validator';
 import { Type } from 'class-transformer';
 import { ApiProperty } from '@nestjs/swagger';
-import { ALIAS_SEGMENT_ERROR, ALIAS_SEGMENT_REGEX } from '../schemas/blended-fields-config.schemas';
+import { ALIAS_SEGMENT_ERROR, ALIAS_SEGMENT_REGEX } from '../schemas/blended-fields-config.schema';
 import { JoinConditionApiDto } from './create-relationship-request-api.dto';
 
 export class UpdateRelationshipRequestApiDto {
