+++ apps/backend/src/data-marts/dto/schemas/join-condition.schema.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+
+export const JoinConditionSchema = z.object({
+  sourceFieldName: z.string().min(1),
+  targetFieldName: z.string().min(1),
+});
+export type JoinCondition = z.infer<typeof JoinConditionSchema>;
+export const JoinConditionsSchema = z.array(JoinConditionSchema);
