+++ apps/backend/src/data-marts/services/blendable-schema.service.ts
@@ -9,8 +9,8 @@ import { DataMartService } from './data-mart.service';
 import { DataMartSchema } from '../data-storage-types/data-mart-schema.type';
 import { DataMartSchemaFieldStatus } from '../data-storage-types/enums/data-mart-schema-field-status.enum';
 import { isNumericFieldType } from '../data-storage-types/field-type-compatibility';
-import { BlendedFieldsConfig, BlendedSource } from '../dto/schemas/blended-fields-config.schemas';
-import { AggregateFunction } from '../dto/schemas/relationship-schemas';
+import { BlendedFieldsConfig, BlendedSource } from '../dto/schemas/blended-fields-config.schema';
+import { AggregateFunction } from '../dto/schemas/aggregate-function.schema';
 import { BusinessViolationException } from '../../common/exceptions/business-violation.exception';
 
 const MAX_TRANSITIVE_DEPTH = 10;
