+++ apps/backend/src/data-marts/services/blended-report-data.service.spec.ts
@@ -634,41 +634,4 @@ describe('BlendedReportDataService', () => {
       expect(context!.chains[1].relationship.id).toBe('rel-bc');
     });
   });
-
-  describe('logBlendedSqlIfNeeded', () => {
-    const makeLogger = () => ({
-      log: jest.fn(),
-      error: jest.fn(),
-      asArrays: jest.fn().mockReturnValue({ logs: [], errors: [] }),
-    });
-
-    it('does nothing when logger is undefined', () => {
-      expect(() =>
-        service.logBlendedSqlIfNeeded({ needsBlending: true, blendedSql: 'SELECT 1' }, undefined)
-      ).not.toThrow();
-    });
-
-    it('does nothing when needsBlending is false', () => {
-      const logger = makeLogger();
-      service.logBlendedSqlIfNeeded({ needsBlending: false }, logger);
-      expect(logger.log).not.toHaveBeenCalled();
-    });
-
-    it('does nothing when blendedSql is missing even though needsBlending is true', () => {
-      const logger = makeLogger();
-      service.logBlendedSqlIfNeeded({ needsBlending: true }, logger);
-      expect(logger.log).not.toHaveBeenCalled();
-    });
-
-    it('logs blended SQL with structured payload when required', () => {
-      const logger = makeLogger();
-      const sql = 'WITH cte AS (SELECT 1) SELECT * FROM cte';
-      service.logBlendedSqlIfNeeded({ needsBlending: true, blendedSql: sql }, logger);
-      expect(logger.log).toHaveBeenCalledWith({
-        type: 'joined-data-marts-sql',
-        message: 'SQL over joined Data Marts used for report execution',
-        sql,
-      });
-    });
-  });
 });
